member, category, product service command/query 분리
