+++ dogdog/src/main/java/com/dogworld/dogdog/category/application/CategoryCommandService.java
@@ -1,52 +1,26 @@
 package com.dogworld.dogdog.category.application;
 
+import com.dogworld.dogdog.category.domain.Category;
+import com.dogworld.dogdog.category.domain.repository.CategoryRepository;
 import com.dogworld.dogdog.category.interfaces.dto.request.CategoryRequest;
 import com.dogworld.dogdog.category.interfaces.dto.response.CategoryCreateResponse;
 import com.dogworld.dogdog.category.interfaces.dto.response.CategoryResponse;
-import com.dogworld.dogdog.category.domain.Category;
-import com.dogworld.dogdog.category.infrastructure.dto.FlatCategoryDto;
-import com.dogworld.dogdog.category.domain.repository.CategoryRepository;
-import com.dogworld.dogdog.product.domain.repository.ProductRepository;
 import com.dogworld.dogdog.global.error.code.ErrorCode;
 import com.dogworld.dogdog.global.error.exception.CustomException;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import com.dogworld.dogdog.product.domain.repository.ProductRepository;
 import java.util.Objects;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class CategoryService {
+@Transactional
+public class CategoryCommandService {
 
   private final CategoryRepository categoryRepository;
   private final ProductRepository productRepository;
 
-  public List<CategoryResponse> getAllCategories() {
-    List<Category> categories = categoryRepository.findAll();
-
-    return categories.stream()
-        .map(CategoryResponse::from)
-        .collect(Collectors.toList());
-  }
-
-  public List<CategoryResponse> getAllCategoriesHierarchy() {
-    List<FlatCategoryDto> flatList = categoryRepository.findAllFlat();
-
-    Map<Long, CategoryResponse> map = buildCategoryMap(flatList);
-    List<CategoryResponse> roots = buildCategoryTree(flatList, map);
-    sortTree(roots);
-
-    return roots;
-  }
-
-  @Transactional
   public CategoryCreateResponse createCategory(CategoryRequest request) {
 
     Category parent = getParentCategory(request.getParentId());
@@ -56,7 +30,6 @@ public class CategoryService {
     return CategoryCreateResponse.from(savedCategory);
   }
 
-  @Transactional
   public CategoryResponse updateCategory(Long categoryId, CategoryRequest request) {
     validateNotSelfAsParent(categoryId, request.getParentId());
     Category category = getCategory(categoryId);
@@ -65,48 +38,10 @@ public class CategoryService {
     return CategoryResponse.from(category);
   }
 
-  @Transactional
   public void deleteCategory(Long categoryId) {
     Category category = getCategory(categoryId);
     validateDeletable(categoryId);
     categoryRepository.delete(category);
-  }
-
-  private Map<Long, CategoryResponse> buildCategoryMap(List<FlatCategoryDto> flatList) {
-    Map<Long, CategoryResponse> map = new HashMap<>();
-
-    for(FlatCategoryDto flatCategoryDto : flatList) {
-      CategoryResponse response = CategoryResponse.from(flatCategoryDto);
-      map.put(flatCategoryDto.getId(), response);
-    }
-    return map;
-  }
-
-  private static List<CategoryResponse> buildCategoryTree(List<FlatCategoryDto> flatList,
-                                                            Map<Long, CategoryResponse> map) {
-    List<CategoryResponse> roots = new ArrayList<>();
-
-    for(FlatCategoryDto dto : flatList) {
-      CategoryResponse child = map.get(dto.getId());
-      Long parentId = dto.getParentId();
-
-      if(parentId == null) {
-        roots.add(child);
-      } else {
-        CategoryResponse parent = map.get(parentId);
-        if(parent != null) {
-          parent.getChildren().add(child);
-        }
-      }
-    }
-    return roots;
-  }
-
-  private void sortTree(List<CategoryResponse> nodes) {
-    nodes.sort(Comparator.comparingInt(CategoryResponse::getSortOrder));
-    for(CategoryResponse node : nodes) {
-      sortTree(node.getChildren());
-    }
   }
 
   private Category getCategory(Long categoryId) {
@@ -139,6 +74,3 @@ public class CategoryService {
     }
   }
 }
-
-
-
