+++ dogdog/src/main/java/com/dogworld/dogdog/category/interfaces/CategoryController.java
@@ -1,7 +1,7 @@
 package com.dogworld.dogdog.category.interfaces;
 
+import com.dogworld.dogdog.category.application.CategoryQueryService;
 import com.dogworld.dogdog.category.interfaces.dto.response.CategoryResponse;
-import com.dogworld.dogdog.category.application.CategoryService;
 import com.dogworld.dogdog.global.common.response.ApiResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -15,17 +15,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/categories")
 public class CategoryController {
 
-  private final CategoryService categoryService;
+  private final CategoryQueryService categoryQueryService;
 
   @GetMapping
   public ResponseEntity<ApiResponse<List<CategoryResponse>>> getAllCategories() {
-    List<CategoryResponse> response = categoryService.getAllCategories();
+    List<CategoryResponse> response = categoryQueryService.getAllCategories();
     return ResponseEntity.ok(ApiResponse.success(response));
   }
 
   @GetMapping("/hierarchy")
   public ResponseEntity<ApiResponse<List<CategoryResponse>>> getAllCategoriesHierarchy() {
-    List<CategoryResponse> response = categoryService.getAllCategoriesHierarchy();
+    List<CategoryResponse> response = categoryQueryService.getAllCategoriesHierarchy();
     return ResponseEntity.ok(ApiResponse.success(response));
   }
 }
