+++ dogdog/src/main/java/com/dogworld/dogdog/member/application/MemberQueryService.java
@@ -1,0 +1,12 @@
+package com.dogworld.dogdog.member.application;
+
+import com.dogworld.dogdog.member.domain.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberQueryService {
+  private final MemberRepository memberRepository;
+
+}
