+++ dogdog/src/main/java/com/dogworld/dogdog/member/interfaces/MemberController.java
@@ -1,8 +1,8 @@
 package com.dogworld.dogdog.member.interfaces;
 
+import com.dogworld.dogdog.member.application.MemberCommandService;
 import com.dogworld.dogdog.member.interfaces.dto.request.MemberRequest;
 import com.dogworld.dogdog.member.interfaces.dto.response.MemberResponse;
-import com.dogworld.dogdog.member.application.MemberService;
 import com.dogworld.dogdog.global.common.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,11 +16,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class MemberController {
 
-  private final MemberService memberService;
+  private final MemberCommandService memberCommandService;
 
   @PostMapping
   public ResponseEntity<ApiResponse<MemberResponse>> createMember(@RequestBody MemberRequest request) {
-    MemberResponse response = memberService.createMember(request);
+    MemberResponse response = memberCommandService.createMember(request);
     return ResponseEntity.ok(ApiResponse.success(response));
   }
 }
