+++ dogdog/src/main/java/com/dogworld/dogdog/product/application/ProductCommandService.java
@@ -1,48 +1,26 @@
 package com.dogworld.dogdog.product.application;
 
-import com.dogworld.dogdog.product.interfaces.dto.request.ProductRequest;
-import com.dogworld.dogdog.product.interfaces.dto.request.ProductSearchCondition;
-import com.dogworld.dogdog.product.interfaces.dto.response.ProductResponse;
 import com.dogworld.dogdog.category.domain.Category;
 import com.dogworld.dogdog.category.domain.repository.CategoryRepository;
-import com.dogworld.dogdog.product.domain.Product;
-import com.dogworld.dogdog.product.infrastructure.ProductQueryRepository;
-import com.dogworld.dogdog.product.domain.repository.ProductRepository;
-import com.dogworld.dogdog.global.common.QueryDslConfig;
 import com.dogworld.dogdog.global.error.code.ErrorCode;
 import com.dogworld.dogdog.global.error.exception.CustomException;
-import java.util.List;
-import java.util.stream.Collectors;
+import com.dogworld.dogdog.product.domain.Product;
+import com.dogworld.dogdog.product.domain.repository.ProductRepository;
+import com.dogworld.dogdog.product.infrastructure.ProductQueryRepository;
+import com.dogworld.dogdog.product.interfaces.dto.request.ProductRequest;
+import com.dogworld.dogdog.product.interfaces.dto.response.ProductResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional(readOnly = true)
+@Transactional
 @RequiredArgsConstructor
-public class ProductService {
-
+public class ProductCommandService {
   private final ProductRepository productRepository;
   private final ProductQueryRepository productQueryRepository;
   private final CategoryRepository categoryRepository;
-  private final QueryDslConfig queryDslConfig;
 
-  public List<ProductResponse> getAllProducts() {
-    List<Product> products = productRepository.findAll();
-
-    return products.stream()
-        .map(ProductResponse::from)
-        .collect(Collectors.toList());
-  }
-
-  public Page<ProductResponse> searchProducts(ProductSearchCondition condition, Pageable pageable) {
-    Page<Product> search = productQueryRepository.search(condition, pageable);
-    return search.map(ProductResponse::from);
-  }
-
-  @Transactional
   public ProductResponse createProduct(ProductRequest request) {
     validateDuplicateName(request.getName());
 
@@ -52,12 +30,6 @@ public class ProductService {
     return ProductResponse.from(savedProduct);
   }
 
-  public ProductResponse getProductById(Long productId) {
-    Product product = getProduct(productId);
-    return ProductResponse.from(product);
-  }
-
-  @Transactional
   public ProductResponse updateProduct(Long productId, ProductRequest request) {
     Product product = getProduct(productId);
     Category category = getCategory(request.getCategoryId());
@@ -65,7 +37,6 @@ public class ProductService {
     return ProductResponse.from(product);
   }
 
-  @Transactional
   public void deleteProduct(Long productId) {
     Product product = getProduct(productId);
 
@@ -90,6 +61,5 @@ public class ProductService {
     return categoryRepository.findById(categoryId)
         .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_CATEGORY));
   }
-
 
 }
