+++ dogdog/src/main/java/com/dogworld/dogdog/product/interfaces/ProductAdminController.java
@@ -1,8 +1,8 @@
 package com.dogworld.dogdog.product.interfaces;
 
+import com.dogworld.dogdog.product.application.ProductCommandService;
 import com.dogworld.dogdog.product.interfaces.dto.request.ProductRequest;
 import com.dogworld.dogdog.product.interfaces.dto.response.ProductResponse;
-import com.dogworld.dogdog.product.application.ProductService;
 import com.dogworld.dogdog.global.common.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -20,26 +20,24 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/admin/products")
 public class ProductAdminController {
 
-  private final ProductService productService;
+  private final ProductCommandService productCommandService;
 
   @PostMapping
   public ResponseEntity<ApiResponse<ProductResponse>> createProduct(@Valid @RequestBody ProductRequest request) {
-    ProductResponse response =  productService.createProduct(request);
+    ProductResponse response =  productCommandService.createProduct(request);
     return ResponseEntity.ok(ApiResponse.success(response));
   }
 
   @PutMapping("/{productId}")
   public ResponseEntity<ApiResponse<ProductResponse>> updateProduct(@PathVariable Long productId,
                                                   @Valid @RequestBody ProductRequest request) {
-    ProductResponse response = productService.updateProduct(productId, request);
+    ProductResponse response = productCommandService.updateProduct(productId, request);
     return ResponseEntity.ok(ApiResponse.success(response));
   }
 
   @DeleteMapping("/{productId}")
   public ResponseEntity<ApiResponse<Void>> deleteProduct(@PathVariable Long productId) {
-    productService.deleteProduct(productId);
+    productCommandService.deleteProduct(productId);
     return ResponseEntity.ok(ApiResponse.success());
   }
-
-
 }
